feat(user-interaction): Implement BookmarkController for GET /users/{userId}/bookmarks #879
