+++ itachallenge-user/src/main/java/com/itachallenge/user/controller/UserController.java
@@ -239,6 +239,8 @@ public class UserController {
                 });
     }
 
+
+
     @Operation(
             summary = "Delete Challenge from User Favorite Challenges",
             description = "Deletes challenge from user favorites",
@@ -344,34 +346,6 @@ public class UserController {
                         log.info("No change, User's '{}' bookmarks doesn't contain Challenge '{}'", userId, challengeId);
                         return ResponseEntity.ok().body(false);
                     }
-                });
-    }
-
-    @Operation(
-            summary = "Gets challenges marked as bookmarks by a user",
-            description = "Returns a set of challenge IDs that the specified user has marked as bookmarked",
-            parameters = {
-                    @Parameter(
-                            name = "userId",
-                            description = "UUID of the user",
-                            required = true,
-                            in = ParameterIn.PATH
-                    )
-            },
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "Set of bookmarked challengeIds by user"),
-                    @ApiResponse(responseCode = "404", description = "User not found"),
-                    @ApiResponse(responseCode = "400", description = "The provided IDs are not valid."),
-                    @ApiResponse(responseCode = "500", description = "Unexpected error")
-            }
-    )
-
-    @GetMapping("/users/{userId}/bookmarks")
-    public Mono<ResponseEntity<Set<UUID>>> getUserBookmarks(@PathVariable String userId) {
-        return bookmarkService.getUserBookmarks(userId)
-                .map(bookmarks -> {
-                    log.info("Retrieved {} bookmarked challenges for user {}", bookmarks.size(), userId);
-                    return ResponseEntity.ok().body(bookmarks);
                 });
     }
